requirements.txt
absl-py==0.13.0
actionlib==1.12.1
angles==1.9.12
asttokens==2.0.5
bondpy==1.8.5
cachetools==4.2.2
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
controller-manager==0.18.4
controller-manager-msgs==0.18.4
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
defusedxml==0.7.1
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
distro==1.6.0
docutils==0.17.1
dynamic-reconfigure==1.6.3
empy==3.3.4
executing==0.8.0
fasteners==0.16.3
future==0.18.2
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
glfw==2.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
icecream==2.1.1
idna==3.2
image-geometry==1.13.0
imageio==2.9.0
importlib-metadata==4.6.4
interactive-markers==1.11.5
joblib==1.0.1
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
kiwisolver==1.3.1
laser-geometry==1.6.7
Markdown==3.3.4
matplotlib==3.3.4
message-filters==1.14.11
moveit-core==1.0.8
msgpack==1.0.2
mujoco-py==2.0.2.8
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1626681920064/work
oauthlib==3.1.1
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
Pygments==2.10.0
pyparsing==2.4.7
PySide2==5.15.2
python-dateutil==2.8.2
python-qt-binding==0.4.3
pytz==2021.1
PyYAML==5.4.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.26.0
requests-oauthlib==1.3.0
resource-retriever==1.12.6
rllab==0.1.0
rosbag==1.14.11
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosdistro==0.8.3
rosgraph==1.14.11
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospkg==1.3.0
rospy==1.14.11
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.7.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
rviz==1.13.18
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sensor-msgs==1.12.7
shiboken2==5.15.2
six==1.16.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
srdfdom==0.5.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
threadpoolctl==2.2.0
topic-tools==1.14.11
torch==1.5.0
typing-extensions==3.10.0.0
urdfdom-py==0.4.5
urllib3==1.26.6
vcstools==0.1.42
Werkzeug==2.0.1
wstool==0.1.17
xacro==1.13.12
zipp==3.5.0
